refactor(bindings,orchestrator): extract config/app_info/update services and split live_state reducer
tauri-bindings:
- Move config, app_info, and update logic out of fat command modules
  into dedicated service submodules under src/services/. Command
  handlers now delegate, leaving them thin and testable.

orchestrator:
- Replace the single live_state/reducer.rs file with a focused module:
  dispatcher routes events; permission, text, and tool submodules each
  own their event-class handling; shared houses common helpers.
  Behaviour is preserved end-to-end.
