+++ crates/tracepilot-orchestrator/src/bridge/live_state/reducer/mod.rs
@@ -1,0 +1,11 @@
+mod dispatcher;
+mod permission;
+mod shared;
+mod text;
+mod tool;
+
+pub use dispatcher::apply_event;
+// Re-exported for the in-crate test module
+// (`bridge::live_state::reducer::MAX_PARTIAL_RESULT_CHARS`).
+#[cfg(test)]
+pub(super) use tool::MAX_PARTIAL_RESULT_CHARS;
