+++ crates/tracepilot-tauri-bindings/src/commands/mod.rs
@@ -4,7 +4,6 @@ pub mod analytics;
 pub mod analytics_executor;
 pub mod blocking_helper;
 pub mod config_cmds;
-mod config_paths;
 pub mod export_import;
 pub mod file_browser;
 pub mod logging;
