+++ crates/tracepilot-tauri-bindings/src/lib.rs
@@ -18,6 +18,7 @@ pub mod error;
 pub mod events;
 mod helpers;
 pub mod ipc_command_names;
+mod services;
 pub mod types;
 mod validators;
